Remove `openshift start node`
