+++ pkg/cmd/server/kubernetes/node/doc.go
@@ -1,0 +1,1 @@
+package node
