+++ pkg/cmd/server/start/network_args.go
@@ -26,8 +26,8 @@ func BindNetworkArgs(args *NetworkArgs, flags *pflag.FlagSet, prefix string) {
 	flags.StringVar(&args.ServiceNetworkCIDR, prefix+"portal-net", args.ServiceNetworkCIDR, "The CIDR string representing the network that portal/service IPs will be assigned from. This must not overlap with any IP ranges assigned to nodes for pods.")
 }
 
-// NewDefaultNetworkArgs returns a new set of network arguments
-func NewDefaultNetworkArgs() *NetworkArgs {
+// NewDefaultMasterNetworkArgs returns a new set of network arguments
+func NewDefaultMasterNetworkArgs() *NetworkArgs {
 	config := &NetworkArgs{
 		NetworkPluginName:  "",
 		ClusterNetworkCIDR: "10.128.0.0/14",
